scripts/requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
openai>=1.0.0